+++ angular/projects/lib/src/lib/types.ts
@@ -1,5 +1,5 @@
 /**
- * gridstack-item.component.ts 12.0.0-dev
+ * gridstack-item.component.ts 12.1.1
  * Copyright (c) 2025 Alain Dumesny - see GridStack root license
  */