+++ src/dd-manager.ts
@@ -1,5 +1,5 @@
 /**
- * dd-manager.ts 12.0.0-dev
+ * dd-manager.ts 12.1.1
  * Copyright (c) 2021-2024 Alain Dumesny - see GridStack root license
  */